requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.0
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
python-dotenv==1.0.0
gunicorn==21.2.0
